% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{BuildLandmarkObject}
\alias{BuildLandmarkObject}
\title{Build a landmark object with correlation analysis and UMAP embedding}
\usage{
BuildLandmarkObject(
  object = NULL,
  sample.obj = NULL,
  rename.sample = NULL,
  nn.name = "weighted.nn",
  weighted.nn.name = NULL,
  landmark.assay.name = "LANDMARK",
  features.to.test = NULL,
  k.nn = 20,
  min.dist = 0.3,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{sample.obj}{Sample-level Seurat object}

\item{rename.sample}{Optional renaming for samples}

\item{nn.name}{Name of the neighbor object to use}

\item{weighted.nn.name}{Deprecated. Use \code{nn.name} instead.}

\item{landmark.assay.name}{Name of landmark assay}

\item{features.to.test}{Features to test for correlation}

\item{k.nn}{Number of nearest neighbors. Default is 20.}

\item{min.dist}{Minimum distance for UMAP. Default is 0.3.}

\item{...}{Additional arguments}
}
\value{
Landmark object with UMAP and correlation data
}
\description{
Build a landmark object with correlation analysis and UMAP embedding
}
\concept{visualization}
