% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{cppls_ondisk}
\alias{cppls_ondisk}
\title{CPLS (Canonical PLS) for On-Disk Matrices}
\usage{
cppls_ondisk(X, ...)
}
\arguments{
\item{X}{Predictor matrix (or an IterableMatrix).}

\item{...}{Arguments passed to methods.}
}
\description{
Generic function for CPLS. The \code{IterableMatrix} method computes
CPLS for BPCells on-disk matrices.
}
