% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{kernelpls}
\alias{kernelpls}
\title{Kernel Partial Least Squares}
\usage{
kernelpls(X, ...)
}
\arguments{
\item{X}{Predictor matrix (or an IterableMatrix).}

\item{...}{Arguments passed to methods.}
}
\description{
Generic function for kernel-PLS. The \code{IterableMatrix} method computes
kernel-PLS for BPCells on-disk matrices.
}
