% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PlotLandmarkObject}
\alias{PlotLandmarkObject}
\title{Visualize landmark-trajectory correlations with UMAP plots}
\usage{
PlotLandmarkObject(
  landmark_obj,
  order = TRUE,
  pt.size = 0.5,
  alpha = 0.8,
  plot.featureplot = FALSE,
  ncol = NULL
)
}
\arguments{
\item{landmark_obj}{Landmark object from BuildLandmarkObject}

\item{order}{Whether to order points by expression}

\item{pt.size}{Point size for plots}

\item{alpha}{Point transparency}

\item{plot.featureplot}{Whether to create feature plots}

\item{ncol}{Number of columns for faceting}
}
\value{
List of ggplot objects
}
\description{
Visualize landmark-trajectory correlations with UMAP plots
}
\concept{visualization}
