% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{RunAndProjectUMAP}
\alias{RunAndProjectUMAP}
\title{Generate a UMAP for a sketched assay and then project it to the full data.}
\usage{
RunAndProjectUMAP(
  object,
  nn.name = NULL,
  k.nn = 5,
  landmark.assay.name = "LANDMARK",
  reduction.name = "sketched.umap",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{a Seurat object}

\item{nn.name}{the name of the Neighbor object used to generate the UMAP.}

\item{k.nn}{the number of neighbors used to genearte the UMAP. If it exceeds the number of k in the nn.name object, will be ignored.}

\item{landmark.assay.name}{the name of the sketched assay}

\item{reduction.name}{the name of the UMAP to be stored; Note that the projected UMAP will be named as "proj.'reduction.name'".}

\item{verbose}{Print progress status}

\item{...}{other parameters that are accepted by RunUMAP() or ProjectUMAP()}
}
\value{
return a Seurat object that contains a UMAP of the sketched assay and a UMAP of the full data
}
\description{
This is a wrapper function that builds on the ProjectUMAP() function, which allows users to
generate UMAP for a subset of the data (e.g., cells from a sketched assay), and then project this
UMAP to any other data.
}
\concept{visualization}
