% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{RunDiffusionMap}
\alias{RunDiffusionMap}
\alias{RunDiffusionMap.default}
\alias{RunDiffusionMap.Assay}
\alias{RunDiffusionMap.StdAssay}
\alias{RunDiffusionMap.Seurat}
\title{Run Diffusion Map}
\usage{
RunDiffusionMap(object, ...)

\method{RunDiffusionMap}{default}(
  object,
  assay = NULL,
  metric = "cosine",
  distance.matrix = NULL,
  ncomp = 20,
  reduction.name = "DiffMap",
  reduction.key = "DC_",
  weight.by.var = TRUE,
  verbose = TRUE,
  seed.use = 101,
  ...
)

\method{RunDiffusionMap}{Assay}(
  object,
  assay = NULL,
  features = NULL,
  ncomp = 20,
  metric = "cosine",
  distance.matrix = NULL,
  reduction.name = "DiffMap",
  reduction.key = "DC_",
  weight.by.var = TRUE,
  verbose = TRUE,
  seed.use = 101,
  ...
)

\method{RunDiffusionMap}{StdAssay}(
  object,
  assay = NULL,
  features = NULL,
  layer = "scale.data",
  ncomp = 20,
  metric = "cosine",
  distance.matrix = NULL,
  reduction.name = "DiffMap",
  reduction.key = "DC_",
  weight.by.var = TRUE,
  verbose = TRUE,
  seed.use = 101,
  ...
)

\method{RunDiffusionMap}{Seurat}(
  object,
  assay = NULL,
  features = NULL,
  layer = "scale.data",
  ncomp = 20,
  metric = "cosine",
  distance.matrix = NULL,
  reduction.name = "DiffMap",
  reduction.key = "DC_",
  weight.by.var = TRUE,
  verbose = TRUE,
  seed.use = 101,
  ...
)
}
\arguments{
\item{object}{An object to run diffusion map on}

\item{...}{Additional arguments to be passed to DiffusionMap()}

\item{assay}{Name of Assay diffusion map is being run on}

\item{metric}{This determines the choice of metric used to measure
distance in the input space. Check DiffusionMap() for details. If a distance.matrix is provided,
this will be ignored.}

\item{distance.matrix}{If set, runs DiffusionMap on the given distance matrix
instead of data matrix.}

\item{ncomp}{Number of components to compute}

\item{reduction.name}{the name of the DimReduc object}

\item{reduction.key}{dimensional reduction key, specifies the string before
the number for the dimension names.}

\item{weight.by.var}{Weight the cell embeddings by the variance of each PC}

\item{verbose}{To print diagnostic messages. Default is TRUE.}

\item{seed.use}{Set a random seed.  Setting NULL will not set a seed.}

\item{features}{Features to compute Diffusion Map}

\item{layer}{The layer in \code{assay} to use when running diffusion map analysis.}
}
\value{
Returns a DimReduc object with diffusion map results
}
\description{
Run diffusion map dimensionality reduction on single-cell data.
}
\examples{
\dontrun{
# Run diffusion map on a Seurat object
seurat_obj <- RunDiffusionMap(seurat_obj, ncomp = 10)
}

}
\concept{dimensional_reduction}
