% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{spls_ondisk}
\alias{spls_ondisk}
\title{Sparse Partial Least Squares for On-Disk Matrices}
\usage{
spls_ondisk(X, ...)
}
\arguments{
\item{X}{Predictor matrix (or an IterableMatrix).}

\item{...}{Arguments passed to methods.}
}
\description{
Generic function for sparse PLS. The \code{IterableMatrix} method computes
sparse PLS for BPCells on-disk matrices using a hybrid R/C++ approach.
}
