% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{TrajDETest}
\alias{TrajDETest}
\alias{TrajDETest.default}
\alias{TrajDETest.Assay}
\alias{TrajDETest.StdAssay}
\alias{TrajDETest.Seurat}
\title{Trajectory Differential Expression Test}
\usage{
TrajDETest(object, ...)

\method{TrajDETest}{default}(
  object,
  traj.var = NULL,
  latent.vars = NULL,
  features = NULL,
  fc.results = NULL,
  verbose = TRUE,
  ...
)

\method{TrajDETest}{Assay}(
  object,
  layer = "counts",
  traj.var = NULL,
  latent.vars = NULL,
  samples = NULL,
  features = NULL,
  logfc.threshold = 0,
  prob.break.point = c(1/3, 2/3),
  min.pct = 0.1,
  min.count = 10,
  pseudocount.use = 1,
  verbose = TRUE,
  ...
)

\method{TrajDETest}{StdAssay}(
  object,
  layer = "counts",
  traj.var = NULL,
  latent.vars = NULL,
  samples = NULL,
  features = NULL,
  logfc.threshold = 0,
  prob.break.point = c(1/3, 2/3),
  min.pct = 0.1,
  min.count = 10,
  pseudocount.use = 1,
  verbose = TRUE,
  ...
)

\method{TrajDETest}{Seurat}(
  object,
  assay = NULL,
  layer = "counts",
  traj.var = NULL,
  latent.vars = NULL,
  samples = NULL,
  features = NULL,
  logfc.threshold = 0,
  prob.break.point = c(1/3, 2/3),
  min.pct = 0.1,
  min.count = 10,
  pseudocount.use = 1,
  complete.resutls = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Expression data matrix, Assay, StdAssay, or Seurat object}

\item{...}{additional parameters to be passed to glmGamPoi::glm_gp or filterByExpr}

\item{traj.var}{a data frame containing the trajectory variable of the samples to be tested against}

\item{latent.vars}{a data frame containing the latent variables (e.g., covariates that might affect the gene expression) to include in regression.}

\item{features}{Genes to test. Default is to use all genes (after QC)}

\item{fc.results}{fc.results calculated by FoldChange(); if not null QC will be performed based on this}

\item{verbose}{Print a progress bar once expression testing begins}

\item{layer}{the data layer to be used for trajectory DE test. Currently only 'counts' is supported.}

\item{samples}{the cells/samples to be included in the DE test}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least X-fold difference (log-scale)
between the top and bottom groups of samples (see pro.break.point for details). Default is 0 (i.e., no filtering).}

\item{prob.break.point}{a numeric vector of probability break points with values in (0, 1). It will be used to calculate 2 quantiles
along the trajectory and to calculate the logfc.}

\item{min.pct}{only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations.}

\item{min.count}{Minimum count threshold for gene filtering}

\item{pseudocount.use}{Pseudocount to add to averaged expression values when calculating logFC. 1 by default.}

\item{assay}{the assay to be used (for Seurat objects)}

\item{complete.resutls}{Whether to return complete results or summary}
}
\value{
A data frame with differential expression results
}
\description{
Perform trajectory-based differential expression analysis using negative binomial regression.
This function tests for genes that change expression along a continuous trajectory variable.
}
\examples{
\dontrun{
# Run trajectory DE test on a Seurat object
de_results <- TrajDETest(seurat_obj, traj.var = "pseudotime")
}

}
\concept{scSLIDE_DE}
