% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{cellanova_calc_BE}
\alias{cellanova_calc_BE}
\title{Perform CellAnova Batch correction}
\usage{
cellanova_calc_BE(
  object = NULL,
  assay = NULL,
  layer = "scale.data",
  integrate_key = NULL,
  features = NULL,
  control_dict = NULL,
  reduction = NULL,
  var_cutoff = 0.9,
  k_max = 1500,
  k_select = NULL,
  new.assay.name = "CORRECTED",
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{the name of the assay to perform the CellAnova correction.}

\item{layer}{the name of the layer to be used to correct for the batch effect. Should be scale.data.}

\item{integrate_key}{A string indicating the smallest batch unit in the meta-data (e.g., library, donor, etc.),
which will be used for integration later.}

\item{features}{features to compute corrected expression for. Defaults to the variable features set in the assay specified.}

\item{control_dict}{A list indicating the control-group assignment of the controls. The name of each element in the
list should correspond to the batch name in the 'integrate_key' column.}

\item{reduction}{the name of the DimReduc object we use as the integrated embeddings. Should be from methods like
Harmony or Seurat-integration methods (e.g., CCA).}

\item{var_cutoff}{the fraction of explained variance to determine the optimal value of k in truncated SVD when calculating
the basis of the batch effect. Default is 0.9.}

\item{k_max}{the maximum of singular values and vectors to compute.}

\item{k_select}{the user-defined number of singular values and vectors to compute (override var_cutoff and k_max).
Default is NULL.}

\item{new.assay.name}{the name for the new assay to store the corrected expression matrix}

\item{verbose}{display progress + messages}
}
\value{
Returns a Seurat object with a new assay added containing the batch-corrected expression matrix
}
\description{
CellAnova (Zhaojun Zhang et al, 2024 Nat Biotech) is a new method that can remove/mitigate batch effect in single-cell data
and return users with a "corrected" data matrix (instead of just a low-dim embedding). Here we implement the 2nd step of the CellAnova.
Specifically, this function is an R-implemention of the CellAnova python package's calc_BE() function
(https://github.com/Janezjz/cellanova).
}
\details{
This function takes a Seurat object and its pre-computed integrated embedding from
methods like Harmony or Seurat-CCA, a batch-effect index, and a case-control index as input, to
estimate the batch effect from the control samples, and correct for it from the full original
expression data.
Most of the procedures are kept the same, with the following modifications:
\itemize{
\item currently we only support one control group.
\item we have additionally implemented a future_lapply() and a more efficient regression framework to enhance the efficiency.\cr
\item the procedure can be done to a "sketched" data and later project to the whole data (for the purposes of efficiency and data balance).\cr
}
}
