% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{FindmmNN}
\alias{FindmmNN}
\title{Find Multi-Modal Nearest Neighbors}
\usage{
FindmmNN(
  object,
  sketch.assay = NULL,
  reduction.list,
  dims.list,
  k.nn = 20,
  knn.range = 200,
  l2.norm = TRUE,
  fix.wnn.weights = NULL,
  weighted.nn.name = "weighted.nn",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{sketch.assay}{The name of the sketched assay of the landmark cells.}

\item{reduction.list}{A list of dimensional reductions, one for each modality}

\item{dims.list}{A list containing the dimensions for each reduction to use}

\item{k.nn}{The number of nearest neighbors to compute for each modality}

\item{knn.range}{Range parameter for nearest neighbor search}

\item{l2.norm}{Perform L2 normalization on the cell embeddings}

\item{fix.wnn.weights}{Pre-specified modality weights. If provided, skips the calculation
and uses these weights directly. Should be a list with the same length as reduction.list.}

\item{weighted.nn.name}{Multimodal neighbor object name}

\item{verbose}{Print progress bars and output}

\item{...}{Arguments passed to other methods}
}
\value{
return a Seurat object that contains a weighted.nn Neighbor object between the landmark cells and all the other cells.
}
\description{
A wrapper function that finds Multi-Modal weights and generate a Multi-Modal NN object.
}
\details{
This function is essentially a wrapper function that perform the following 2 procedures:
\itemize{
\item Seurat:::FindModalityWeights
\item Seurat:::MultiModalNN
}
}
\concept{multimodal}
