% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_level_object.R
\name{GenerateSampleObject}
\alias{GenerateSampleObject}
\title{Generate a sample-level count matrix based on landmark assay and its weighted.nn object}
\usage{
GenerateSampleObject(
  object,
  nn.name = NULL,
  k.nn = 5,
  sketch.assay = "LANDMARK",
  return.seurat = TRUE,
  new_assay_name = "LMC",
  group.by = "ident",
  normalization.method = "ChiSquared",
  scale.factor = 10000,
  rename.group.by = NULL,
  add.meta.data = TRUE,
  cells.use = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{nn.name}{Name of the Neighbor object to use for the calculation
(e.g. \code{"weighted.nn"} as produced by \code{\link{FindmmNN}} or
\code{\link{PrepareSampleObject}}).}

\item{k.nn}{the number of nearest neighbors to perform the summing}

\item{sketch.assay}{the name of the sketch.assay you used to perform the FindmmNN()}

\item{return.seurat}{Whether to return the data as a Seurat object. Default is TRUE}

\item{new_assay_name}{Name for the new assay containing landmark counts}

\item{group.by}{A single meta-data column to group cells by (e.g. Donor ID); 'ident' by default.
To group on a combination of columns, add a merged column to the object beforehand and pass
that, e.g. \code{object$batch_rep <- paste(object$batch, object$replicate, sep = "_")}.}

\item{normalization.method}{Method for normalization. Supports LogNormalize and ChiSquared. see details at
\code{\link{NormalizeData}} and \code{\link{NormalizeChiSquared}}}

\item{scale.factor}{Scale factor for Log-Normalization, see \code{\link{NormalizeData}}}

\item{rename.group.by}{if rename.group.by is NULL, the rownames of the landmark matrix will used the original cell IDs.
But, user can indicate which meta-data column to use to rename the rows.
A suffix of "_LM" + order number will be added automatically.}

\item{add.meta.data}{if TRUE, the function will automatically detect sample-level meta-data (based on 'group.by') and append it
to the sample-level object; if FALSE, it will not do so.}

\item{cells.use}{An optional character vector of cell names to subset before building
the sample-level matrix. When non-NULL, only cells in this vector (that are also present
in the NN object) are retained. Default is NULL (use all cells).}

\item{verbose}{Print progress and diagnostic messages}

\item{...}{Arguments to be passed to methods such as \code{\link{CreateSeuratObject}}}
}
\value{
return a Seurat object that contains a count matrix with number of landmark as rows, sample as columns
}
\description{
Returns summed counts of weighted.nn for each landmark cell within each sample.
}
\concept{scSLIDE}
