% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{RunPLS}
\alias{RunPLS}
\alias{RunPLS.default}
\alias{RunPLS.IterableMatrix}
\alias{RunPLS.Assay}
\alias{RunPLS.StdAssay}
\alias{RunPLS.Seurat}
\title{Run Partial Least Squares (PLS) on Seurat Objects}
\usage{
RunPLS(object, ...)

\method{RunPLS}{default}(
  object,
  assay = NULL,
  ncomp = 20,
  Y = NULL,
  Y.add = NULL,
  pls.function = c("plsr", "spls", "cppls"),
  verbose = TRUE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.name = "pls",
  reduction.key = "PLS_",
  seed.use = 42,
  eta = 0.5,
  ...
)

\method{RunPLS}{IterableMatrix}(
  object,
  assay = NULL,
  ncomp = 20,
  Y = NULL,
  Y.add = NULL,
  pls.function = c("plsr", "spls", "cppls"),
  verbose = TRUE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.name = "pls",
  reduction.key = "PLS_",
  seed.use = 42,
  eta = 0.5,
  ...
)

\method{RunPLS}{Assay}(
  object,
  assay = NULL,
  features = NULL,
  ncomp = 20,
  Y = NULL,
  Y.add = NULL,
  pls.function = c("plsr", "spls", "cppls"),
  verbose = TRUE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.name = "pls",
  reduction.key = "PLS_",
  seed.use = 42,
  eta = 0.5,
  ...
)

\method{RunPLS}{StdAssay}(
  object,
  assay = NULL,
  features = NULL,
  layer = "scale.data",
  ncomp = 20,
  Y = NULL,
  Y.add = NULL,
  pls.function = c("plsr", "spls", "cppls"),
  verbose = TRUE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.name = "pls",
  reduction.key = "PLS_",
  seed.use = 42,
  eta = 0.5,
  ...
)

\method{RunPLS}{Seurat}(
  object,
  assay = NULL,
  features = NULL,
  ncomp = 20,
  Y = NULL,
  Y.add = NULL,
  pls.function = c("plsr", "spls", "cppls"),
  verbose = TRUE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.name = "pls",
  reduction.key = "PLS_",
  seed.use = 42,
  eta = 0.5,
  ...
)
}
\arguments{
\item{object}{An object to run PLS on}

\item{...}{Additional arguments to be passed to the PLS function}

\item{assay}{Name of Assay PLS is being run on}

\item{ncomp}{Number of components to compute}

\item{Y}{a vector or matrix of responses, i.e., the dependent variable that PLS regresses X on. The length / number of rows should be the same as the number of cells.
Y can have multiple columns.}

\item{Y.add}{a vector or matrix of additional responses containing relevant information about the observations. Only used for cppls.}

\item{pls.function}{PLS function from pls package to run (options: plsr, spls, cppls)}

\item{verbose}{Print the top genes associated with high/low loadings for
the components}

\item{ndims.print}{components to print genes for}

\item{nfeatures.print}{Number of genes to print for each component}

\item{reduction.name}{the name of the DimReduc object}

\item{reduction.key}{dimensional reduction key, specifies the string before
the number for the dimension names.}

\item{seed.use}{Set a random seed.  Setting NULL will not set a seed.}

\item{eta}{Thresholding parameter that controls the sparsity of the spls method (larger --> sparser). eta should be between 0 and 1.}

\item{features}{Features to compute PLS on}

\item{layer}{The layer in \code{assay} to use when running PLS analysis.}
}
\value{
Returns a DimReduc object with PLS results
}
\description{
Performs Partial Least Squares regression analysis on single-cell data.
This function provides methods for plsr, spls, and cppls from the pls and spls packages.
}
\examples{
\dontrun{
# Run PLS on a Seurat object
seurat_obj <- RunPLS(seurat_obj, Y = "condition", ncomp = 10)
}

}
\concept{dimensional_reduction}
