% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SampleLevelDimPlot}
\alias{SampleLevelDimPlot}
\title{Visualize the landmark-trajectory relevance and the sample-level cell density}
\usage{
SampleLevelDimPlot(
  object = NULL,
  sample.obj = NULL,
  rename.sample = NULL,
  weighted.nn.name = "weighted.nn",
  landmark.assay.name = "LANDMARK",
  features.to.test = NULL,
  order = TRUE,
  pt.size = 0.5,
  alpha = 0.8,
  ncol = NULL,
  return.seurat.obj = TRUE,
  k.nn = 20,
  min.dist = 0.3,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{sample.obj}{Sample-level Seurat object}

\item{rename.sample}{Optional sample renaming}

\item{weighted.nn.name}{Name of weighted NN object}

\item{landmark.assay.name}{Name of landmark assay}

\item{features.to.test}{Features to test for correlation}

\item{order}{Whether to order points}

\item{pt.size}{Point size}

\item{alpha}{Point transparency}

\item{ncol}{Number of columns}

\item{return.seurat.obj}{Whether to return Seurat object}

\item{k.nn}{Number of nearest neighbors}

\item{min.dist}{Minimum distance for UMAP}

\item{...}{Additional arguments}
}
\value{
List containing plots and optionally landmark object
}
\description{
Visualize the landmark-trajectory relevance and the sample-level cell density
}
\concept{visualization}
